- Game logic

+++ database/migrations/2022_10_05_220901_create_sentences_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('positive_answers')->default(0)->index();
             $table->integer('negative_answers')->default(0)->index();
             $table->integer('total_answers')->default(0)->index();
-            $table->decimal('word_reliability')->default(0)->index();
+            $table->float('word_reliability')->default(0)->index();
             $table->bigInteger('external_id')->nullable();
             $table->timestamps();
         });
